Add sub-bar fractional positioning, grid snap system, and clone automation fixes
Container.startBar/lengthBars changed from Int to Double for sub-beat precision.
Backward-compatible Codable decoding handles both Int and Double in saved projects.

Grid system additions:
- GridMode enum (adaptive / fixed resolution) with toolbar toggle
- SnapResolution extended with triplet values (1/4T, 1/8T, 1/16T, 1/32T)
- TimelineViewModel.snapToGrid() unifies snap logic across all timeline interactions
- GridOverlayView renders subdivisions based on active grid mode
- Snap-to-grid plumbed through RulerView, ContainerView, TrackLaneView

Clone automation fix:
- Container.copyField(.automation) and resolveInheritedFields now remap
  automation lane containerIDs from parent to clone, so automation targets
  the correct container instance

Also wires onAutomationDataChanged to push live breakpoint edits to the
scheduler during playback without rebuilding the audio graph.

+++ Sources/LoopsApp/ViewModels/ClipboardState.swift
@@ -18,7 +18,7 @@ public final class ClipboardState {
     public var clipboard: [ClipboardContainerEntry] = []
 
     /// The leftmost start bar of copied containers, used for offset calculation on paste.
-    public var clipboardBaseBar: Int = 1
+    public var clipboardBaseBar: Double = 1.0
 
     /// Section region metadata copied with section copy operations.
     public var clipboardSectionRegion: SectionRegion?
